road network test cases finished

+++ env_model/roadNetwork/road_network.h
@@ -52,7 +52,7 @@ class RoadNetwork {
         * @param id lanelet ID
         * @return pointer to lanelet
         */
-        std::shared_ptr<Lanelet> findLaneletById(size_t id);
+        std::shared_ptr<Lanelet> findLaneletById(int id);
 
         /**
         * Given a set of lanes and a polygon shape, finds a single lane which intersect with the shape (interstate)
